+++ src/compileModel.ts
@@ -6,7 +6,7 @@ import type { QueryConstructorOptions } from './Query';
 import Query, { type Filter } from './Query';
 import type Schema from './Schema';
 import type { GenericObject, MvRecord } from './shared/types';
-import { ensureArray } from './shared/utils';
+import { ensureArray } from './utils';
 
 // #region Types
 export interface ModelConstructorOptionsBase {
