+++ src/utils/handleRequiredValidation.ts
@@ -1,4 +1,4 @@
-import type { SchemaValidator, ValidationFunction, Validator } from '../types';
+import type { SchemaValidator, ValidationFunction, Validator } from '../shared/types';
 
 /**
  * Create Validator for required validation
